refactor: split clusters.rs into focused modules
Split src-tauri/src/commands/clusters.rs (883 LOC) into a clusters/
module directory with focused files:

- types.rs: ClusterInfo, ConnectionStatus, HealthCheckResult, NamespaceResult
- kubeconfig.rs: load_kubeconfig_from_sources, build_kubeconfig_for_connect,
  merge_kubeconfig_files, is_self_contained, load_configured_namespaces + tests
- commands.rs: all #[command] functions
- mod.rs: re-exports so external use paths don't change

Pure move, no logic changes. All 14 existing tests pass.

Closes #288
